@@ -1,6 +1,6 @@
 cmake_minimum_required(VERSION 3.9)
 project(KalmanFilter)
-set(CMAKE_CXX_STANDARD 11)
+set(CMAKE_CXX_STANDARD 14)
 
 find_package(PkgConfig)
 pkg_check_modules(GTKMM gtkmm-3.0)
